requirements.txt
jinja2
GitPython
pyyaml